# Remove RemoveDuplicateJoinRestrictions() function