--- a/src/backend/distributed/utils/listutils.c
+++ b/src/backend/distributed/utils/listutils.c
@@ -258,15 +258,3 @@ GenerateListFromElement(void *listElement, int listLength)
 
 	return list;
 }
-
-
-/*
- * LeftListIsSubset returns true if lhs is subset of rhs list. Note that input
- * lists' entries should implement equal() function.
- */
-bool
-LeftListIsSubset(const List *lhs, const List *rhs)
-{
-	List *listDifference = list_difference(lhs, rhs);
-	return list_length(listDifference) == 0;
-}