--- a/src/include/distributed/distributed_planner.h
+++ b/src/include/distributed/distributed_planner.h
@@ -76,7 +76,6 @@ typedef struct JoinRestrictionContext
 {
 	List *joinRestrictionList;
 	bool hasSemiJoin;
-	bool hasOnlyInnerJoin;
 } JoinRestrictionContext;
 
 typedef struct JoinRestriction