RemoveDuplicateJoinRestrictions() function was introduced with the aim of decrasing the overall planning times by eliminating the duplicate JOIN restriction entries (#1989). 

**However, it turns out that the function itself is so CPU intensive with a very high algorithmic complexity, it hurts a lot more than it helps. The function is a clear example of premature optimization.**

The table below shows the difference clearly:

  | distributed query planning      time master | RemoveDuplicateJoinRestrictions()   execution time on master | distributed query planning this PR
-- | -- | -- | --
5 table INNER JOIN | 9 msec | 2msec | 7 msec
10 table INNER JOIN | 227 msec | 194 msec | 29  msec
20 table INNER JOIN | 1 sec 235 msec | 1  sec 139  msec | 90 msecs
50 table INNER JOIN | 24 seconds | 21 seconds | 1.5 seconds
100 table INNER JOIN | 2 minutes 16 secods | 1 minute 53 seconds | 23 seconds
250 table INNER JOIN | Bottleneck on JoinClauseList | 18 minutes 52 seconds | Bottleneck on JoinClauseList
  |   |   |  
  |   |   |  
5 table INNER JOIN in subquery | 9 msec | 0 msec | 6 msec
10 table INNER JOIN subquery | 33 msec | 10 msec | 32 msec
20 table INNER JOIN subquery | 132 msec | 67 msec | 123 msec
50 table INNER JOIN subquery | 1.2  seconds | 900 msec | 500 msec
100 table INNER JOIN subquery | 6 seconds | 5  seconds | 2 seconds
250 table INNER JOIN subquery | 54 seconds | 37 seconds | 20  seconds
  |   |   |  
5 table LEFT JOIN | 5 msec | 0 msec | 5 msec
10 table LEFT JOIN | 11 msec | 0 msec | 13 msec
20 table LEFT JOIN | 26 msec | 2 msec | 30 msec
50 table LEFT JOIN | 150 msec | 15 msec | 193 msec
100 table LEFT JOIN | 757 msec | 71 msec | 722 msec
250 table LEFT JOIN | 8 seconds | 600 msec | 8 seconds
  |   |   |  
5 JOINs among 2 table   JOINs | 37 msec | 11 msec | 25 msec
10 JOINs among 2 table   JOINs | 536 msec | 306 msec | 352 msec
20 JOINs among 2 table   JOINs | 794 msec | 181 msec | 640 msec
50 JOINs among 2 table   JOINs | 25 seconds | 2 seconds | 22 seconds
100 JOINs among 2 table   JOINs | Bottleneck on JoinClauseList | 9 seconds | Bottleneck on JoinClauseList
150 JOINs among 2 table   JOINs | Bottleneck on JoinClauseList | 46 seconds | Bottleneck on JoinClauseList


On top of the performance penalty, the function had a critical bug #4255, and with #4254 we hit one more important bug. It should be fixed by adding the followig check to the ContextCoversJoinRestriction():
```
static bool
JoinRelIdsSame(JoinRestriction *leftRestriction, JoinRestriction *rightRestriction)
{
	Relids leftInnerRelIds = leftRestriction->innerrel->relids;
	Relids rightInnerRelIds = rightRestriction->innerrel->relids;
	if (!bms_equal(leftInnerRelIds, rightInnerRelIds))
	{
		return false;
	}

	Relids leftOuterRelIds = leftRestriction->outerrel->relids;
	Relids rightOuterRelIds = rightRestriction->outerrel->relids;
	if (!bms_equal(leftOuterRelIds, rightOuterRelIds))
	{
		return false;
	}

	return true;
}
```

However, adding this eliminates all the benefits tha RemoveDuplicateJoinRestrictions() brings.

I've used the commands here to generate the JOINs mentioned in the PR: https://gist.github.com/onderkalaci/fe8654f9df5916c7af4c7c5eb892561e#file-gistfile1-txt